a-dashboard/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
aiohttp>=3.9
aiosqlite>=0.20
jinja2>=3.1
python-dotenv>=1.0
